requirements.txt
scrapy
scrapy-fake-useragent
psycopg2-binary
